Removed main.yml
I've come to realise that my goal of having the .apk file on my repository is bad practice, which is why it was so hard to do. The real goal should've been taking the artefact and deploying it, but I don't have Google Play or IOS deployment set up. Will come back to this later!