Created build category folders
The Desktop build of the game will require assets not included in the Mobile build, and vice versa. Thus the files will be split into three folders: /Mobile, /Desktop, and /Core. The /Core directory contains files used in both builds.

Blank .txt files were created so that the folders could be committed.

P.S.: The reason the two builds differ is that, in the desktop build, you'll play as a Phrog who plays the "Phunky Phrogs" mobile build on their Phrog Phone. The phone will take centre screen, with a pretty backdrop behind it. This allows for minimal gameplay changes between builds.